Add filter logic
